#2066: can't cast uintptr_t to 'unsigned long'
git-svn-id: https://xpra.org/svn/Xpra/trunk@21148 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/codecs/csc_libyuv/colorspace_converter.pyx
@@ -127,7 +127,7 @@ class YUVImageWrapper(ImageWrapper):
         return "libyuv.YUVImageWrapper"
 
     def free(self):                             #@DuplicatedSignature
-        log("libyuv.YUVImageWrapper.free() cython_buffer=%#x", <unsigned long> self.cython_buffer)
+        log("libyuv.YUVImageWrapper.free() cython_buffer=%#x", <uintptr_t> self.cython_buffer)
         ImageWrapper.free(self)
         if self.cython_buffer>0:
             free(<void *> (<uintptr_t> self.cython_buffer))
